Ray Train Release Test: Add Image Classification Jpeg
