+++ release/train_tests/benchmark/image_classification/image_classification_jpeg/torch_jpeg_image_iterable_dataset.py
@@ -2,11 +2,9 @@ from typing import Iterator, List, Optional, Tuple
 import time
 import logging
 import io
-import pandas as pd
-import pyarrow.parquet as pq
-import multiprocessing
 import boto3
 from botocore.exceptions import NoCredentialsError
+from PIL import Image
 
 import torch
 from torch.utils.data import IterableDataset
@@ -14,19 +12,9 @@ from torch.utils.data import IterableDataset
 import ray.train
 import ray
 
-from image_classification.imagenet import get_preprocess_map_fn
+from image_classification_jpeg.imagenet import get_preprocess_map_fn
 
 logger = logging.getLogger(__name__)
-
-# Set multiprocessing start method to 'spawn' for CUDA compatibility
-if torch.cuda.is_available():
-    try:
-        multiprocessing.set_start_method("spawn", force=True)
-        logger.info(
-            "[DataLoader] Set multiprocessing start method to 'spawn' for CUDA compatibility"
-        )
-    except RuntimeError:
-        logger.info("[DataLoader] Multiprocessing start method already set")
 
 # AWS configuration
 AWS_REGION = "us-west-2"
@@ -36,42 +24,28 @@ AWS_REGION = "us-west-2"
 
 
 @ray.remote(num_cpus=0.25)
-def _fetch_parquet_metadata(bucket: str, key: str, file_url: str) -> Tuple[str, int]:
-    """Standalone Ray task to fetch Parquet row count using S3 Select.
+def _fetch_file_size(bucket: str, key: str, file_url: str) -> Tuple[str, int]:
+    """Standalone Ray task to fetch file size from S3 using head_object.
 
     Args:
         bucket: S3 bucket name
         key: S3 object key
-        file_url: Full S3 URL for logging
+        file_url: Full S3 URL for logging.
 
     Returns:
-        Tuple of (file_url, row_count)
+        Tuple of (file_url, file_size in bytes)
     """
-    import boto3
-    import logging
-
     logger = logging.getLogger(__name__)
     s3_client = boto3.client("s3", region_name=AWS_REGION)
 
-    # Use S3 Select to count rows
-    query = "SELECT COUNT(*) FROM s3object"
-    response = s3_client.select_object_content(
-        Bucket=bucket,
-        Key=key,
-        Expression=query,
-        ExpressionType="SQL",
-        InputSerialization={"Parquet": {}},
-        OutputSerialization={"CSV": {}},
-    )
-
-    # Parse the response to get row count
-    row_count = 0
-    for event in response["Payload"]:
-        if "Records" in event:
-            row_count = int(event["Records"]["Payload"].decode("utf-8").strip())
-
-    logger.info(f"[DataLoader] File {file_url} has {row_count} rows")
-    return file_url, row_count
+    try:
+        head_response = s3_client.head_object(Bucket=bucket, Key=key)
+        file_size = head_response["ContentLength"]  # File size in bytes
+        logger.info(f"[DataLoader] File {file_url} has size {file_size} bytes")
+        return file_url, file_size
+    except Exception as e:
+        logger.error(f"[DataLoader] Error fetching size for {file_url}: {str(e)}")
+        return file_url, -1  # Indicating failure
 
 
 class S3Reader:
@@ -124,14 +98,14 @@ class S3Reader:
     def _collect_file_info(
         self, bucket: str, prefix: str
     ) -> Tuple[List[str], List[int]]:
-        """Collect file URLs and their row counts in parallel using Ray tasks.
+        """Collect file URLs and their sizes in parallel using Ray tasks.
 
         Args:
             bucket: S3 bucket name
             prefix: S3 prefix to list
 
         Returns:
-            Tuple of (file_urls, row_counts)
+            Tuple of (file_urls, file_sizes_in_bytes)
         """
         response = self.s3_client.list_objects_v2(Bucket=bucket, Prefix=prefix)
 
@@ -142,7 +116,7 @@ class S3Reader:
         tasks = []
         for file in response["Contents"]:
             file_url = f"s3://{bucket}/{file['Key']}"
-            task = _fetch_parquet_metadata.remote(bucket, file["Key"], file_url)
+            task = _fetch_file_size.remote(bucket, file["Key"], file_url)
             tasks.append(task)
 
         # Wait for all tasks to complete
@@ -153,12 +127,12 @@ class S3Reader:
         results = ray.get(tasks)
 
         # Unzip results
-        file_urls, file_rows = zip(*results) if results else ([], [])
+        file_urls, file_size_bytes = zip(*results) if results else ([], [])
 
         logger.info(
             f"[DataLoader] Worker {worker_rank}: Collected metadata for {len(file_urls)} files"
         )
-        return list(file_urls), list(file_rows)
+        return list(file_urls), list(file_size_bytes)
 
     def _distribute_files(
         self,
@@ -167,58 +141,58 @@ class S3Reader:
         worker_rank: int,
         num_workers: int,
     ) -> List[str]:
-        """Distribute files among workers based on row counts.
+        """Distribute files among workers based on file sizes.
 
         Args:
             file_urls: List of file URLs
-            file_rows: List of row counts per file
+            file_rows: List of file sizes in bytes
             worker_rank: Current worker rank
             num_workers: Total number of workers
 
         Returns:
             List of file URLs assigned to this worker
         """
-        # Sort files by row count (descending) for better distribution
-        files_with_rows = sorted(
+        # Sort files by size (descending) for better distribution
+        files_with_sizes = sorted(
             zip(file_urls, file_rows), key=lambda x: x[1], reverse=True
         )
-        file_urls = [f[0] for f in files_with_rows]
-        file_rows = [f[1] for f in files_with_rows]
+        file_urls = [f[0] for f in files_with_sizes]
+        file_rows = [f[1] for f in files_with_sizes]
 
         # If no workers or files, return all files
         if num_workers <= 1 or not file_urls:
             logger.info(
                 f"[DataLoader] Worker {worker_rank}: Single worker or no files, returning all {len(file_urls)} files "
-                f"with total {sum(file_rows)} rows"
+                f"with total {sum(file_rows)} bytes"
             )
             return file_urls
 
-        # Calculate target rows per worker for balanced distribution
-        total_rows = sum(file_rows)
-        target_rows_per_worker = total_rows / num_workers
+        # Calculate target size per worker for balanced distribution
+        total_size = sum(file_rows)
+        target_size_per_worker = total_size / num_workers
         logger.info(
-            f"[DataLoader] Worker {worker_rank}: Total rows: {total_rows}, Target rows per worker: {target_rows_per_worker:.0f}"
+            f"[DataLoader] Worker {worker_rank}: Total size: {total_size} bytes, Target size per worker: {target_size_per_worker:.0f} bytes"
         )
 
-        # Distribute files to minimize row count differences between workers
+        # Distribute files to minimize size differences between workers
         worker_files = [[] for _ in range(num_workers)]
-        worker_rows = [0] * num_workers
+        worker_sizes = [0] * num_workers
 
-        # Assign files to workers, trying to keep row counts balanced
-        for file_url, row_count in zip(file_urls, file_rows):
-            # Find worker with minimum current row count
-            min_rows_worker = min(range(num_workers), key=lambda w: worker_rows[w])
-            worker_files[min_rows_worker].append(file_url)
-            worker_rows[min_rows_worker] += row_count
+        # Assign files to workers, trying to keep sizes balanced
+        for file_url, file_size in zip(file_urls, file_rows):
+            # Find worker with minimum current size
+            min_size_worker = min(range(num_workers), key=lambda w: worker_sizes[w])
+            worker_files[min_size_worker].append(file_url)
+            worker_sizes[min_size_worker] += file_size
 
         # Get this worker's files
         my_files = worker_files[worker_rank]
-        my_row_count = worker_rows[worker_rank]
+        my_size = worker_sizes[worker_rank]
 
         logger.info(
             f"[DataLoader] Worker {worker_rank}: Assigned {len(my_files)}/{len(file_urls)} files "
-            f"with {my_row_count}/{total_rows} rows "
-            f"({my_row_count/total_rows*100:.1f}% of total)"
+            f"with {my_size}/{total_size} bytes "
+            f"({my_size/total_size*100:.1f}% of total)"
         )
         return my_files
 
@@ -270,10 +244,10 @@ class S3Reader:
             logger.info(
                 "[DataLoader] Collecting file metadata for balanced distribution"
             )
-            file_urls, file_rows = self._collect_file_info(bucket, prefix)
+            file_urls, file_size_bytes = self._collect_file_info(bucket, prefix)
             return self._distribute_files(
                 file_urls=file_urls,
-                file_rows=file_rows,
+                file_rows=file_size_bytes,
                 worker_rank=worker_rank,
                 num_workers=num_workers,
             )
@@ -286,11 +260,11 @@ class S3Reader:
             raise self.S3FileError(f"Error listing files from {url}: {str(e)}")
 
 
-class S3ParquetImageIterableDataset(S3Reader, IterableDataset):
+class S3JpegImageIterableDataset(S3Reader, IterableDataset):
     """An iterable dataset that loads images from S3-stored Parquet files.
 
     This dataset:
-    1. Reads Parquet files from S3 one row group at a time
+    1. Reads Jpeg files from S3
     2. Processes images with optional random transforms
     3. Yields (image, label) tensors
     4. Supports row limits per worker for controlled data processing
@@ -316,50 +290,23 @@ class S3ParquetImageIterableDataset(S3Reader, IterableDataset):
         self.limit_rows_per_worker = limit_rows_per_worker
         self.random_transforms = random_transforms
 
-    def _read_parquet_file(self, file_url: str) -> Iterator[pd.DataFrame]:
-        """Read a Parquet file from S3 one row group at a time.
+    def _fetch_image(self, file_url: str) -> Image:
+        """Fetch and decode JPEG image from S3.
 
         Args:
-            file_url: S3 URL of the Parquet file
+            file_url: S3 URL of the image file
 
-        Yields:
-            DataFrame containing one row group at a time
+        Returns:
+            PIL Image object
         """
         try:
-            start_time = time.time()
-            worker_info = torch.utils.data.get_worker_info()
-            worker_id = worker_info.id if worker_info else 0
-            logger.info(
-                f"[Dataset] Worker {worker_id}: Reading Parquet file: {file_url}"
-            )
-
-            # Get parquet file metadata
             bucket, key = self._parse_s3_url(file_url)
             response = self.s3_client.get_object(Bucket=bucket, Key=key)
-            parquet_file = pq.ParquetFile(io.BytesIO(response["Body"].read()))
-            num_row_groups = parquet_file.num_row_groups
-
-            logger.info(
-                f"[Dataset] Worker {worker_id}: Found {num_row_groups} row groups in {file_url}"
-            )
-
-            for row_group in range(num_row_groups):
-                # Read row group and convert to pandas
-                table = parquet_file.read_row_group(row_group)
-                df = table.to_pandas()
-                yield df
-
-            total_time = time.time() - start_time
-            logger.info(
-                f"[Dataset] Worker {worker_id}: Completed reading {file_url} in {total_time:.2f}s"
-            )
-
+            image_data = response["Body"].read()
+            image = Image.open(io.BytesIO(image_data))
+            return image
         except Exception as e:
-            worker_info = torch.utils.data.get_worker_info()
-            worker_id = worker_info.id if worker_info else 0
-            logger.error(
-                f"[Dataset] Worker {worker_id}: Error reading file {file_url}: {str(e)}"
-            )
+            logger.error(f"Error fetching image from {file_url}: {str(e)}")
             raise
 
     def __iter__(self) -> Iterator[Tuple[torch.Tensor, torch.Tensor]]:
@@ -411,56 +358,37 @@ class S3ParquetImageIterableDataset(S3Reader, IterableDataset):
                     )
                     break
 
-                for df in self._read_parquet_file(file_url):
-                    # Skip DataFrame if we've reached the limit
-                    if (
-                        self.limit_rows_per_worker is not None
-                        and rows_processed >= self.limit_rows_per_worker
-                    ):
-                        break
+                try:
+                    # Fetch image from S3
+                    image = self._fetch_image(file_url)
 
-                    for _, row in df.iterrows():
-                        # Skip row if we've reached the limit
-                        if (
-                            self.limit_rows_per_worker is not None
-                            and rows_processed >= self.limit_rows_per_worker
-                        ):
-                            break
+                    # Process image and convert to tensors
+                    processed = preprocess_fn(image)
+                    image = torch.as_tensor(processed["image"], dtype=torch.float32)
+                    label = torch.as_tensor(processed["label"], dtype=torch.int64)
 
-                        try:
-                            # Process row and convert to tensors
-                            processed = preprocess_fn(row)
-                            image = torch.as_tensor(
-                                processed["image"], dtype=torch.float32
-                            )
-                            label = torch.as_tensor(
-                                processed["label"], dtype=torch.int64
-                            )
+                    rows_processed += 1
 
-                            rows_processed += 1
+                    # Log progress periodically
+                    if rows_processed % self.LOG_FREQUENCY == 0:
+                        current_time = time.time()
+                        elapsed_time = current_time - last_log_time
+                        rows_per_second = (
+                            self.LOG_FREQUENCY / elapsed_time if elapsed_time > 0 else 0
+                        )
+                        logger.info(
+                            f"[Dataset] Worker {worker_id}: "
+                            f"Processed {rows_processed} rows ({rows_per_second:.2f} rows/sec)"
+                        )
+                        last_log_time = current_time
 
-                            # Log progress periodically
-                            if rows_processed % self.LOG_FREQUENCY == 0:
-                                current_time = time.time()
-                                elapsed_time = current_time - last_log_time
-                                rows_per_second = (
-                                    self.LOG_FREQUENCY / elapsed_time
-                                    if elapsed_time > 0
-                                    else 0
-                                )
-                                logger.info(
-                                    f"[Dataset] Worker {worker_id}: "
-                                    f"Processed {rows_processed} rows ({rows_per_second:.2f} rows/sec)"
-                                )
-                                last_log_time = current_time
+                    yield image, label
 
-                            yield image, label
-
-                        except Exception as e:
-                            logger.error(
-                                f"[Dataset] Worker {worker_id}: Error processing row: {str(e)}"
-                            )
-                            continue
+                except Exception as e:
+                    logger.error(
+                        f"[Dataset] Worker {worker_id}: Error processing row: {str(e)}"
+                    )
+                    continue
 
             # Log final statistics
             total_time = time.time() - total_start_time
