+++ release/train_tests/benchmark/image_classification/imagenet.py
@@ -1,9 +1,4 @@
-import io
-
-import numpy as np
-from PIL import Image
 import torchvision
-from torchvision.transforms.functional import pil_to_tensor
 
 
 def _get_sysnet_mapping():
@@ -533,263 +528,6 @@ def _get_sysnet_mapping():
         "n03131574": "crib, cot",
         "n03133878": "Crock Pot",
         "n03134739": "croquet ball",
-        "n03141823": "crutch",
-        "n03146219": "cuirass",
-        "n03160309": "dam, dike, dyke",
-        "n03179701": "desk",
-        "n03180011": "desktop computer",
-        "n03187595": "dial telephone, dial phone",
-        "n03188531": "diaper, nappy, napkin",
-        "n03196217": "digital clock",
-        "n03197337": "digital watch",
-        "n03201208": "dining table, board",
-        "n03207743": "dishrag, dishcloth",
-        "n03207941": "dishwasher, dish washer, dishwashing machine",
-        "n03208938": "disk brake, disc brake",
-        "n03216828": "dock, dockage, docking facility",
-        "n03218198": "dogsled, dog sled, dog sleigh",
-        "n03220513": "dome",
-        "n03223299": "doormat, welcome mat",
-        "n03240683": "drilling platform, offshore rig",
-        "n03249569": "drum, membranophone, tympan",
-        "n03250847": "drumstick",
-        "n03255030": "dumbbell",
-        "n03259280": "Dutch oven",
-        "n03271574": "electric fan, blower",
-        "n03272010": "electric guitar",
-        "n03272562": "electric locomotive",
-        "n03290653": "entertainment center",
-        "n03291819": "envelope",
-        "n03297495": "espresso maker",
-        "n03314780": "face powder",
-        "n03325584": "feather boa, boa",
-        "n03337140": "file, file cabinet, filing cabinet",
-        "n03344393": "fireboat",
-        "n03345487": "fire engine, fire truck",
-        "n03347037": "fire screen, fireguard",
-        "n03355925": "flagpole, flagstaff",
-        "n03372029": "flute, transverse flute",
-        "n03376595": "folding chair",
-        "n03379051": "football helmet",
-        "n03384352": "forklift",
-        "n03388043": "fountain",
-        "n03388183": "fountain pen",
-        "n03388549": "four-poster",
-        "n03393912": "freight car",
-        "n03394916": "French horn, horn",
-        "n03400231": "frying pan, frypan, skillet",
-        "n03404251": "fur coat",
-        "n03417042": "garbage truck, dustcart",
-        "n03424325": "gasmask, respirator, gas helmet",
-        "n03425413": "gas pump, gasoline pump, petrol pump, island dispenser",
-        "n03443371": "goblet",
-        "n03444034": "go-kart",
-        "n03445777": "golf ball",
-        "n03445924": "golfcart, golf cart",
-        "n03447447": "gondola",
-        "n03447721": "gong, tam-tam",
-        "n03450230": "gown",
-        "n03452741": "grand piano, grand",
-        "n03457902": "greenhouse, nursery, glasshouse",
-        "n03459775": "grille, radiator grille",
-        "n03461385": "grocery store, grocery, food market, market",
-        "n03467068": "guillotine",
-        "n03476684": "hair slide",
-        "n03476991": "hair spray",
-        "n03478589": "half track",
-        "n03481172": "hammer",
-        "n03482405": "hamper",
-        "n03483316": "hand blower, blow dryer, blow drier, hair dryer, hair drier",
-        "n03485407": "hand-held computer, hand-held microcomputer",
-        "n03485794": "handkerchief, hankie, hanky, hankey",
-        "n03492542": "hard disc, hard disk, fixed disk",
-        "n03494278": "harmonica, mouth organ, harp, mouth harp",
-        "n03495258": "harp",
-        "n03496892": "harvester, reaper",
-        "n03498962": "hatchet",
-        "n03527444": "holster",
-        "n03529860": "home theater, home theatre",
-        "n03530642": "honeycomb",
-        "n03532672": "hook, claw",
-        "n03534580": "hoopskirt, crinoline",
-        "n03535780": "horizontal bar, high bar",
-        "n03538406": "horse cart, horse-cart",
-        "n03544143": "hourglass",
-        "n03584254": "iPod",
-        "n03584829": "iron, smoothing iron",
-        "n03590841": "jack-o'-lantern",
-        "n03594734": "jean, blue jean, denim",
-        "n03594945": "jeep, landrover",
-        "n03595614": "jersey, T-shirt, tee shirt",
-        "n03598930": "jigsaw puzzle",
-        "n03599486": "jinrikisha, ricksha, rickshaw",
-        "n03602883": "joystick",
-        "n03617480": "kimono",
-        "n03623198": "knee pad",
-        "n03627232": "knot",
-        "n03630383": "lab coat, laboratory coat",
-        "n03633091": "ladle",
-        "n03637318": "lampshade, lamp shade",
-        "n03642806": "laptop, laptop computer",
-        "n03649909": "lawn mower, mower",
-        "n03657121": "lens cap, lens cover",
-        "n03658185": "letter opener, paper knife, paperknife",
-        "n03661043": "library",
-        "n03662601": "lifeboat",
-        "n03666591": "lighter, light, igniter, ignitor",
-        "n03670208": "limousine, limo",
-        "n03673027": "liner, ocean liner",
-        "n03676483": "lipstick, lip rouge",
-        "n03680355": "Loafer",
-        "n03690938": "lotion",
-        "n03691459": "loudspeaker, speaker, speaker unit, loudspeaker system, speaker system",  # noqa: E501
-        "n03692522": "loupe, jeweler's loupe",
-        "n03697007": "lumbermill, sawmill",
-        "n03706229": "magnetic compass",
-        "n03709823": "mailbag, postbag",
-        "n03710193": "mailbox, letter box",
-        "n03710637": "maillot",
-        "n03710721": "maillot, tank suit",
-        "n03717622": "manhole cover",
-        "n03720891": "maraca",
-        "n03721384": "marimba, xylophone",
-        "n03724870": "mask",
-        "n03729826": "matchstick",
-        "n03733131": "maypole",
-        "n03733281": "maze, labyrinth",
-        "n03733805": "measuring cup",
-        "n03742115": "medicine chest, medicine cabinet",
-        "n03743016": "megalith, megalithic structure",
-        "n03759954": "microphone, mike",
-        "n03761084": "microwave, microwave oven",
-        "n03763968": "military uniform",
-        "n03764736": "milk can",
-        "n03769881": "minibus",
-        "n03770439": "miniskirt, mini",
-        "n03770679": "minivan",
-        "n03773504": "missile",
-        "n03775071": "mitten",
-        "n03775546": "mixing bowl",
-        "n03776460": "mobile home, manufactured home",
-        "n03777568": "Model T",
-        "n03777754": "modem",
-        "n03781244": "monastery",
-        "n03782006": "monitor",
-        "n03785016": "moped",
-        "n03786901": "mortar",
-        "n03787032": "mortarboard",
-        "n03788195": "mosque",
-        "n03788365": "mosquito net",
-        "n03791053": "motor scooter, scooter",
-        "n03792782": "mountain bike, all-terrain bike, off-roader",
-        "n03792972": "mountain tent",
-        "n03793489": "mouse, computer mouse",
-        "n03794056": "mousetrap",
-        "n03796401": "moving van",
-        "n03803284": "muzzle",
-        "n03804744": "nail",
-        "n03814639": "neck brace",
-        "n03814906": "necklace",
-        "n03825788": "nipple",
-        "n03832673": "notebook, notebook computer",
-        "n03837869": "obelisk",
-        "n03838899": "oboe, hautboy, hautbois",
-        "n03840681": "ocarina, sweet potato",
-        "n03841143": "odometer, hodometer, mileometer, milometer",
-        "n03843555": "oil filter",
-        "n03854065": "organ, pipe organ",
-        "n03857828": "oscilloscope, scope, cathode-ray oscilloscope, CRO",
-        "n03866082": "overskirt",
-        "n03868242": "oxcart",
-        "n03868863": "oxygen mask",
-        "n03871628": "packet",
-        "n03873416": "paddle, boat paddle",
-        "n03874293": "paddlewheel, paddle wheel",
-        "n03874599": "padlock",
-        "n03876231": "paintbrush",
-        "n03877472": "pajama, pyjama, pj's, jammies",
-        "n03877845": "palace",
-        "n03884397": "panpipe, pandean pipe, syrinx",
-        "n03887697": "paper towel",
-        "n03888257": "parachute, chute",
-        "n03888605": "parallel bars, bars",
-        "n03891251": "park bench",
-        "n03891332": "parking meter",
-        "n03895866": "passenger car, coach, carriage",
-        "n03899768": "patio, terrace",
-        "n03902125": "pay-phone, pay-station",
-        "n03903868": "pedestal, plinth, footstall",
-        "n03908618": "pencil box, pencil case",
-        "n03908714": "pencil sharpener",
-        "n03916031": "perfume, essence",
-        "n03920288": "Petri dish",
-        "n03924679": "photocopier",
-        "n03929660": "pick, plectrum, plectron",
-        "n03929855": "pickelhaube",
-        "n03930313": "picket fence, paling",
-        "n03930630": "pickup, pickup truck",
-        "n03933933": "pier",
-        "n03935335": "piggy bank, penny bank",
-        "n03937543": "pill bottle",
-        "n03938244": "pillow",
-        "n03942813": "ping-pong ball",
-        "n03944341": "pinwheel",
-        "n03947888": "pirate, pirate ship",
-        "n03950228": "pitcher, ewer",
-        "n03954731": "plane, carpenter's plane, woodworking plane",
-        "n03956157": "planetarium",
-        "n03958227": "plastic bag",
-        "n03961711": "plate rack",
-        "n03967562": "plow, plough",
-        "n03970156": "plunger, plumber's helper",
-        "n03976467": "Polaroid camera, Polaroid Land camera",
-        "n03976657": "pole",
-        "n03977966": "police van, police wagon, paddy wagon, patrol wagon, wagon, black Maria",  # noqa: E501
-        "n03980874": "poncho",
-        "n03982430": "pool table, billiard table, snooker table",
-        "n03983396": "pop bottle, soda bottle",
-        "n03991062": "pot, flowerpot",
-        "n03992509": "potter's wheel",
-        "n03995372": "power drill",
-        "n03998194": "prayer rug, prayer mat",
-        "n04004767": "printer",
-        "n04005630": "prison, prison house",
-        "n04008634": "projectile, missile",
-        "n04009552": "projector",
-        "n04019541": "puck, hockey puck",
-        "n04023962": "punching bag, punch bag, punching ball, punchball",
-        "n04026417": "purse",
-        "n04033901": "quill, quill pen",
-        "n04033995": "quilt, comforter, comfort, puff",
-        "n04037443": "racer, race car, racing car",
-        "n04039381": "racket, racquet",
-        "n04040759": "radiator",
-        "n04041544": "radio, wireless",
-        "n04044716": "radio telescope, radio reflector",
-        "n04049303": "rain barrel",
-        "n04065272": "recreational vehicle, RV, R.V.",
-        "n04067472": "reel",
-        "n04069434": "reflex camera",
-        "n04070727": "refrigerator, icebox",
-        "n04074963": "remote control, remote",
-        "n04081281": "restaurant, eating house, eating place, eatery",
-        "n04086273": "revolver, six-gun, six-shooter",
-        "n04090263": "rifle",
-        "n04099969": "rocking chair, rocker",
-        "n04111531": "rotisserie",
-        "n04116512": "rubber eraser, rubber, pencil eraser",
-        "n04118538": "rugby ball",
-        "n04118776": "rule, ruler",
-        "n04120489": "running shoe",
-        "n04125021": "safe",
-        "n04127249": "safety pin",
-        "n04131690": "saltshaker, salt shaker",
-        "n04133789": "sandal",
-        "n04136333": "sarong",
-        "n04141076": "sax, saxophone",
-        "n04141327": "scabbard",
-        "n04141975": "scale, weighing machine",
-        "n04146614": "school bus",
         "n04147183": "schooner",
         "n04149813": "scoreboard",
         "n04152593": "screen, CRT screen",
@@ -1017,14 +755,6 @@ IMAGENET_WNID_TO_LABEL = _get_sysnet_mapping()
 SORTED_WNIDS = sorted(IMAGENET_WNID_TO_LABEL.keys())
 IMAGENET_WNID_TO_ID = {wnid: SORTED_WNIDS.index(wnid) for wnid in SORTED_WNIDS}
 
-IMAGENET_PARQUET_SPLIT_S3_ROOT = (
-    "s3://ray-benchmark-data-internal/imagenet/parquet_split"
-)
-IMAGENET_PARQUET_SPLIT_S3_DIRS = {
-    split: f"{IMAGENET_PARQUET_SPLIT_S3_ROOT}/{split}"
-    for split in ["train", "val", "test"]
-}
-
 
 def get_transform(to_torch_tensor: bool = True, random_transforms: bool = True):
     transforms = []
@@ -1055,31 +785,3 @@ def get_transform(to_torch_tensor: bool = True, random_transforms: bool = True):
     ]
 
     return torchvision.transforms.Compose(transforms)
-
-
-def get_preprocess_map_fn(decode_image: bool = True, random_transforms: bool = True):
-    """Get a map function that transforms a row of the dataset to the format
-    expected by the training loop.
-
-    The output is a dict with 2 keys:
-    - "image": np.array of the transformed, normalized image
-    - "label": An integer index of the WNID
-
-    TODO: Move these preprocessing utilities shared between tests to a separate file.
-    """
-    crop_resize_transform = get_transform(
-        to_torch_tensor=False, random_transforms=random_transforms
-    )
-
-    def map_fn(row):
-        assert "image" in row and "label" in row, row.keys()
-
-        if decode_image:
-            row["image"] = pil_to_tensor(Image.open(io.BytesIO(row["image"]))) / 255.0
-
-        row["image"] = np.array(crop_resize_transform(row["image"]))
-        row["label"] = IMAGENET_WNID_TO_ID[row["label"]]
-
-        return {"image": row["image"], "label": row["label"]}
-
-    return map_fn
